pyproject.toml
[project]
name = "stdbench"
version = "0.1.0"
description = "Python library designed for benchmarking standard C++ parallel algorithms, algotirhm policies on different standard libraries implementations, compilers and architectures."
authors = [
    {name = "Stanislav Sidelnikov", email = "sidelnikov.si.001@gmail.com"},
]
dependencies = [
    "conan==2.8.1",
    "flake8==7.1.1",
    "black==24.10.0",
    "pytest==8.3.4",
    "pyyaml==6.0.2"
]
requires-python = "==3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.black]
line_length = "120"

[tool.flake8]
show_source = true 
statistics = true

